Added footer
It isn't actually a footer at the moment, I haven't learned how to add one properly yet, but it's getting closer. The socials buttons will eventually be replaced with icons instead of text, too.

I've also reformatted a lot of stuff, exporting the stylesheet (which should reduce copied code for the other pages) and putting buttons inside forms so that they can act as links without the need for javascript code.